backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pillow==10.1.0
rembg==2.0.50
opencv-python-headless==4.8.1.78
boto3==1.29.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
numpy==1.24.3